Typo and some questions
Typo, plus some other changes I can't fix.

Index 5 and Index 6 are both Yang.  It sounds like Index 5 should maybe be Farah, but I'm not certain.

Index 12 is Darius, but Index 13 responds as if it should be Farah talking in Index 12.

The LoreYenLoWang file is missing from HyadesRim\descriptions, and so the hyperlink in the conversation doesn't show up.