Typos
Also, something is weird with the knockdown of these Yellow Bird Cicadas.  They are becoming Unsteady when expected, but then are knocked down when the stability bar fills up the second to last box.  The last box is completely empty when it gets knocked down.

One more thing, I destroyed all of the waves and then the bird base once the last wave deployed, but the mission didn't end properly.  I had to use the debug console to end it.