darkForest_3_story Typos
The Newgrange only has 50 hit points.  Instead of destroying the 250 hit point pumping station, I just targeted the dropship 'building' directly and completed the mission that way.  It would be much better if the Newgrange had like 2500 hit points or something large like that.

Also, I don't understand why Adept Smith is shown firing a gunshot when the Unknown commander is killed.  The Newgrange is already destroyed by that point.