Update comancheMoon_1_simpleBattle.json
The Mech in the first slot was not the mech that was locked into the contract.  It was the Mech in the second slot.  I'm wondering if it's a matter of referencing Mech Bay 0 instead of Mech Bay 1, but I don't see anything in this file that would alter that.